telegram_checkin_bot/requirements.txt
python-telegram-bot==20.6
pandas
openpyxl
APScheduler
